programming-logic-evaluation/src/4-list-available-30-minute-slots/list-available-30-minute-slots.ts
// Importa os tipos necessários do arquivo '../types'
import { CalendarAvailability, CalendarEvent, CalendarSlot } from '../types';

// Define a função listAvailable30MinuteSlots que retorna slots de 30 minutos disponíveis
export const listAvailable30MinuteSlots = (
  availability: CalendarAvailability, // Disponibilidade do calendário
  events: Array<CalendarEvent>, // Array de eventos agendados
  range: [Date, Date], // Intervalo de datas a ser considerado
): Array<CalendarSlot> => {
  // Inicializa um array para armazenar os slots disponíveis
  const availableSlots: Array<CalendarSlot> = [];

  // Seu código vai aqui para calcular os slots disponíveis...

  // Retorna a lista de slots disponíveis
  return [];
};

/* 

Explicação do Código
Importação de Tipos:

O código começa importando três tipos (CalendarAvailability, CalendarEvent, CalendarSlot) do arquivo '../types'. Esses tipos são usados para garantir que os dados passados para a função tenham a estrutura correta.
Definição da Função:

A função listAvailable30MinuteSlots é exportada, o que permite que ela seja utilizada em outros módulos. Ela aceita três parâmetros:
availability: Representa os períodos de tempo em que o calendário está disponível para agendamentos.
events: Um array de eventos já agendados, que precisam ser considerados ao determinar a disponibilidade.
range: Um par de datas que define o intervalo de tempo no qual estamos buscando slots disponíveis.
Array de Slots Disponíveis:

Dentro da função, um array chamado availableSlots é inicializado para armazenar os slots de 30 minutos que estão disponíveis após a verificação dos eventos existentes e da disponibilidade.
Lógica para Calcular Slots:

O trecho de código marcado como // Seu código vai aqui é onde a lógica para calcular quais slots de 30 minutos estão disponíveis deve ser implementada. Essa lógica envolveria verificar os períodos de disponibilidade em relação aos eventos agendados.
Retorno:

A função retorna o array availableSlots, que, neste momento, está vazio. O objetivo é que, após a implementação da lógica, ele contenha todos os slots de 30 minutos que não conflitam com os eventos já agendados.
Resumo
Esta função é uma base para calcular os horários disponíveis de 30 minutos em um calendário, considerando a disponibilidade e os eventos já existentes. O código ainda precisa da implementação da lógica para determinar quais slots estão realmente disponíveis.

*/